I have added a joystick for the movment
I have added a joystick so that the player can move in any direction. The speed and rotation speed can be changed if needs be. I have also had a go at spawning random enemies that will spawn in random places in the world. I have also added a simple trail to the player which can be discarded if it doesn't look right. Finally i have created a particle effect that would be shown when the cannon is shot. This can be re done or discarged it it dosen't look right.