Sound
I have added sound to the buttons, in the main menu and the sample scene. So when they are clicked they make a noise. if anybody wants to change the sound, go for it. I just need to fully fix it as when you press the start button for the game window, the sound is played.